requirements.txt
bert-score
torch
transformers
datasets 
scikit-learn
pandas
numpy
tabulate
nltk
python-dotenv
tqdm